services/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sentry-sdk[fastapi]==1.40.5
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
httpx==0.27.0
stripe==8.4.0
twilio==9.0.0
resend==0.8.0
python-dotenv==1.0.1
boto3==1.34.46
tenacity==8.2.3
openai==1.12.0
anthropic==0.18.1
redis==5.0.1
celery==5.3.6
flower==2.0.1
email-validator==2.1.0.post1
reportlab==4.1.0
python-dateutil==2.8.2
